Move EngineExtensions::Decorators to EngineExtensions
This module doesn't just deal with decorators anymore, but also
loads views and controllers, so this seems like a better name.
Separating it into two modules wouldn't be simple because they
are interlinked.
